+++ apps/docs/content/components/input/regex-validation.ts
@@ -3,7 +3,7 @@ const App = `import {Input} from "@nextui-org/react";
 export default function App() {
   const [value, setValue] = React.useState("junior2nextui.org");
 
-  const validateEmail = (value) => value.match(/^[A-Z0-9._%+-]+@[A-Z0-9.-]+.[A-Z]{2,4}$/i);
+  const validateEmail = (value) => value.match(/^[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,4}$/i);
 
   const isInvalid = React.useMemo(() => {
     if (value === "") return false;
@@ -19,7 +19,7 @@ export default function App() {
       variant="bordered"
       isInvalid={isInvalid}
       color={isInvalid ? "danger" : "success"}
-      errorMessage={isInvalid && "Please enter a valid email"}
+      errorMessage="Please enter a valid email"
       onValueChange={setValue}
       className="max-w-xs"
     />