+++ packages/components/modal/src/use-modal.ts
@@ -5,10 +5,15 @@ import {AriaModalOverlayProps} from "@react-aria/overlays";
 import {useAriaModalOverlay} from "@nextui-org/use-aria-modal-overlay";
 import {useCallback, useId, useRef, useState, useMemo, ReactNode} from "react";
 import {modal} from "@nextui-org/theme";
-import {HTMLNextUIProps, mapPropsVariants, PropGetter} from "@nextui-org/system";
+import {
+  HTMLNextUIProps,
+  mapPropsVariants,
+  PropGetter,
+  useProviderContext,
+} from "@nextui-org/system";
 import {useAriaButton} from "@nextui-org/use-aria-button";
 import {useFocusRing} from "@react-aria/focus";
-import {clsx, dataAttr} from "@nextui-org/shared-utils";
+import {clsx, dataAttr, objectToDeps} from "@nextui-org/shared-utils";
 import {ReactRef, useDOMRef} from "@nextui-org/react-utils";
 import {useOverlayTriggerState} from "@react-stately/overlays";
 import {OverlayTriggerProps} from "@react-stately/overlays";
@@ -74,6 +79,8 @@ interface Props extends HTMLNextUIProps<"section"> {
 export type UseModalProps = Props & OverlayTriggerProps & AriaModalOverlayProps & ModalVariantProps;
 
 export function useModal(originalProps: UseModalProps) {
+  const globalContext = useProviderContext();
+
   const [props, variantProps] = mapPropsVariants(originalProps, modal.variantKeys);
 
   const {
@@ -81,7 +88,7 @@ export function useModal(originalProps: UseModalProps) {
     as,
     className,
     classNames,
-    disableAnimation = false,
+
     isOpen,
     defaultOpen,
     onOpenChange,
@@ -104,6 +111,9 @@ export function useModal(originalProps: UseModalProps) {
   const [headerMounted, setHeaderMounted] = useState(false);
   const [bodyMounted, setBodyMounted] = useState(false);
 
+  const disableAnimation =
+    originalProps.disableAnimation ?? globalContext?.disableAnimation ?? false;
+
   const dialogId = useId();
   const headerId = useId();
   const bodyId = useId();
@@ -122,6 +132,7 @@ export function useModal(originalProps: UseModalProps) {
   const {modalProps, underlayProps} = useAriaModalOverlay(
     {
       isDismissable,
+      shouldBlockScroll,
       isKeyboardDismissDisabled,
     },
     state,
@@ -138,8 +149,9 @@ export function useModal(originalProps: UseModalProps) {
     () =>
       modal({
         ...variantProps,
+        disableAnimation,
       }),
-    [...Object.values(variantProps)],
+    [objectToDeps(variantProps), disableAnimation],
   );
 
   const getDialogProps: PropGetter = (props = {}, ref = null) => ({
@@ -157,7 +169,6 @@ export function useModal(originalProps: UseModalProps) {
   const getBackdropProps = useCallback<PropGetter>(
     (props = {}) => ({
       className: slots.backdrop({class: classNames?.backdrop}),
-      onClick: () => state.close(),
       ...underlayProps,
       ...props,
     }),