+++ packages/core/theme/src/components/pagination.ts
@@ -161,7 +161,6 @@ const pagination = tv({
     isCompact: false,
     isDisabled: false,
     showShadow: false,
-    disableAnimation: false,
     disableCursorAnimation: false,
   },
   compoundVariants: [