+++ packages/utilities/aria-utils/src/collections/index.ts
@@ -1,3 +1,7 @@
-export * from "./item";
-export * from "./section";
-export * from "./types";
+export type {ItemProps} from "./item";
+export {BaseItem} from "./item";
+
+export type {SectionProps} from "./section";
+export {BaseSection} from "./section";
+
+export type {CollectionProps, PartialNode} from "./types";